requirements.txt
scipy
asgiref==3.4.1
Django==3.1.7
docx2txt==0.8
hazm==0.7.0
nltk==3.3
patool==1.12
Pillow==8.1.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
django_extensions==3.1.3
PersianStemmer
matplotlib
scikit-learn
requests
django-after-response==0.2.2
ZipUnicode
djangorestframework
mysqlclient
pandas
openpyxl
pdfplumber
multilingual-pdf2text
spacy
#tensorflow
unidecode
jdatetime==4.1.0
gensim
statsmodels
#pylcs
opencv-python==4.6.0.66
# pytesseract==0.3.9
PyPDF2
pytesseract==0.3.3
tqdm==4.42.1
pdf2image==1.12.1
python-docx==0.8.10  
elasticsearch==7.17.4
textract
django-cors-headers
django-ssl-auth
seaborn
networkx
fuzzy-c-means
yellowbrick
# transformers
SentencePiece
protobuf --no-binary=protobuf
pytest-django
num2fawords
transformers